src/main/java/org/hussard/codewars/CuckooClock.java
package org.hussard.codewars;

/**
 * <p>
 * Le coucou sort de l'horloge et sonne une fois au quart d'heure, à la demi-heure et aux trois quarts d'heure. Au début
 * de chaque heure (1-12), il sonne l'heure. Etant donné l'heure actuelle et un nombre n, déterminez l'heure à laquelle
 * le coucou a sonné n fois.
 * </p>
 * <p>
 * Paramètres d'entrée : initial_time - une chaîne au format "HH:MM", où 1 ≤ HH ≤ 12 et 0 ≤ MM ≤ 59, avec des 0 en tête
 * si nécessaire. n - un nombre entier représentant le nombre cible de carillons, avec 1 <= n <= 200.
 * </p>
 * <p>
 * Valeur de retour : L'heure à laquelle le coucou a sonné n fois - une chaîne au même format que initial_time.
 * </p>
 * <p>
 * Si le coucou émet un carillon à l'heure initiale, il faut l'inclure dans le décompte. Si le nème carillon est atteint
 * au cours de l'heure, indiquez l'heure au début de cette heure (c'est-à-dire en supposant que le coucou a terminé son
 * carillon avant que la minute ne soit écoulée).
 * </p>
 * <p>
 * Exemple : <br>"03:38", 19 doit renvoyer "06:00".
 * <br>Explication :
 * <br>Il sonne une fois à "03:45", 4 fois à "04:00", une fois à
 * "04:15", "04:30", "04:45", 5 fois à "05:00", <br>une fois à "05:15", "05:30", "05:45". <br>À ce stade, il a sonné 16 fois, et
 * le 19e carillon se produit lorsqu'il sonne 6 fois à "06:00".
 * </p>
 * <p>
 * <b>Source : International Collegiate Programming Contest, North Central North American Regional, 2023.</b>
 * </p>
 */
public class CuckooClock {
    private CuckooClock() {
        throw new UnsupportedOperationException("CuckooClock is a utility class and cannot be instantiated");
    }

    public static String cuckooClock(String inputTime,
                                     int chimes) {

        // Insert code here
        return inputTime;
    }
}
